+++ src/index.ts
@@ -1,6 +1,7 @@
 export * from './core/types';
 export * from './core/assetCache';
 export * from './core/frameLoader';
+export * from './core/archiveLoader';
 export * from './core/canvasRenderer';
 export * from './core/stage';
 export { nativeAnimationDriver } from './core/nativeDriver';